feat: code quality improvements
- Rate-limit Nominatim geocoding (1.1s interval) with proper User-Agent header
- Sanitize custom theme imports (hex color validation, name/description limits)
- Lazy-load MapLibre GL (~500KB saved on initial load)
- Add OpenStreetMap tile attribution
- Fix duplicate drawPoint calls in export
- Fix Indonesian error messages in export (Gagal → Failed)
- Remove unnecessary dynamic import in drawMarkerToCtx
- Split monolithic form.js (1478 lines) into 5 focused modules
- Add test suite (23 tests) with vitest
