+++ package.json
@@ -5,7 +5,9 @@
 	"scripts": {
 		"dev": "vite",
 		"build": "vite build",
-		"preview": "vite preview"
+		"preview": "vite preview",
+		"test": "vitest run",
+		"test:watch": "vitest"
 	},
 	"dependencies": {
 		"html2canvas": "^1.4.1",
@@ -14,8 +16,10 @@
 	},
 	"devDependencies": {
 		"autoprefixer": "^10.4.24",
+		"jsdom": "^28.1.0",
 		"postcss": "^8.5.6",
 		"tailwindcss": "3.4.1",
-		"vite": "^5.0.0"
+		"vite": "^5.0.0",
+		"vitest": "^4.0.18"
 	}
 }
