+++ src/core/custom-themes.js
@@ -1,5 +1,21 @@
 const STORAGE_KEY = 'map-to-poster:custom-themes';
 
+const HEX_COLOR_RE = /^#[0-9a-fA-F]{3}([0-9a-fA-F]{3})?$/;
+const COLOR_KEYS = ['bg', 'text', 'water', 'parks', 'road_motorway', 'road_primary', 'road_secondary', 'road_tertiary', 'road_residential', 'road_default', 'route'];
+
+function sanitizeTheme(theme) {
+	const clean = { ...theme };
+	if (typeof clean.name === 'string') {
+		clean.name = clean.name.slice(0, 100);
+	}
+	for (const key of COLOR_KEYS) {
+		if (key in clean && !HEX_COLOR_RE.test(clean[key])) {
+			clean[key] = '#000000';
+		}
+	}
+	return clean;
+}
+
 export function loadCustomThemes() {
 	try {
 		const raw = localStorage.getItem(STORAGE_KEY);
@@ -9,7 +25,7 @@ export function loadCustomThemes() {
 
 export function saveCustomTheme(key, theme) {
 	const all = loadCustomThemes();
-	all[key] = theme;
+	all[key] = sanitizeTheme(theme);
 	localStorage.setItem(STORAGE_KEY, JSON.stringify(all));
 }
 
@@ -50,7 +66,7 @@ export function importCustomThemesFromJSON(jsonString) {
 	for (const [key, theme] of Object.entries(parsed)) {
 		if (typeof theme === 'object' && theme !== null && theme.name) {
 			const finalKey = all[key] ? 'custom_' + Date.now() + '_' + imported : key;
-			all[finalKey] = theme;
+			all[finalKey] = sanitizeTheme(theme);
 			imported++;
 		} else {
 			skipped++;
